+++ src/Gateway/ApiGateway/Dockerfile
@@ -1,0 +1,19 @@
+FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
+WORKDIR /app
+EXPOSE 80
+
+FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
+WORKDIR /src
+COPY ["src/Gateway/ApiGateway/ApiGateway.csproj", "src/Gateway/ApiGateway/"]
+RUN dotnet restore "./src/Gateway/ApiGateway/ApiGateway.csproj"
+COPY . .
+WORKDIR "/src/src/Gateway/ApiGateway"
+RUN dotnet build "./ApiGateway.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "./ApiGateway.csproj" -c Release -o /app/publish /p:UseAppHost=false
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "ApiGateway.dll"]
