+++ src/Gateway/ApiGateway/Program.cs
@@ -1,0 +1,47 @@
+using Ocelot.DependencyInjection;
+using Ocelot.Middleware;
+
+namespace GatewayApi
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            new WebHostBuilder()
+            .UseKestrel(options =>
+            {
+                var env = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+                if (env != "Development")
+                {
+                    options.ListenAnyIP(80);
+                }
+
+            })
+            .UseContentRoot(Directory.GetCurrentDirectory())
+            .ConfigureAppConfiguration((hostingContext, config) =>
+            {
+                config
+                    .SetBasePath(hostingContext.HostingEnvironment.ContentRootPath)
+                    .AddJsonFile("appsettings.json", true, true)
+                    .AddJsonFile($"appsettings.{hostingContext.HostingEnvironment.EnvironmentName}.json", true, true)
+                    .AddJsonFile("ocelot.json")
+                    .AddEnvironmentVariables();
+            })
+            .ConfigureServices(s =>
+            {
+                s.AddOcelot();
+            })
+            .ConfigureLogging((hostingContext, logging) =>
+            {
+                //add your logging
+            })
+            .UseIISIntegration()
+            .Configure(app =>
+            {
+                app.UseOcelot().Wait();
+            })
+            .Build()
+            .Run();
+        }
+    }
+}
