--- src/LucasBenHur.MotorcycleMaintenance.Api/Controllers/MotorcycleController.cs
@@ -1,25 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace Api.Controllers
-{
-    [ApiController]
-    [Tags("motos")]
-    [Route("motos")]
-    public class MotorcycleController : ControllerBase
-    {
-        private readonly ILogger<MotorcycleController> _logger;
-
-        public MotorcycleController(
-            ILogger<MotorcycleController> logger)
-        {
-            _logger = logger;
-        }
-
-        [HttpGet]
-        public IActionResult Get()
-        {
-            _logger.LogInformation("Hello, World!");
-            return Ok("Hello, World!");
-        }
-    }
-}
