--- src/LucasBenHur.MotorcycleMaintenance.Api/Program.cs
@@ -1,44 +1,0 @@
-﻿
-using Microsoft.OpenApi.Models;
-
-namespace Api
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var builder = WebApplication.CreateBuilder(args);
-
-            // Add services to the container.
-
-            builder.Services.AddControllers();
-
-            builder.Services.AddEndpointsApiExplorer();
-
-            builder.Services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new OpenApiInfo
-                {
-                    Title = "Sistema de Manutenção de Motos",
-                    Version = "v1"
-                });
-            });
-
-            var app = builder.Build();
-
-            app.UseSwagger();
-            app.UseSwaggerUI();
-
-            app.UseHttpsRedirection();
-
-            app.UseAuthorization();
-
-            app.MapControllers();
-
-            if (!app.Environment.IsDevelopment())
-                app.Run("http://*:80");
-
-            app.Run();
-        }
-    }
-}
