--- src/LucasBenHur.MotorcycleMaintenance.Api/appsettings.Development.json
@@ -1,8 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  }
-}
