--- src/LucasBenHur.MotorcycleMaintenance.Worker/Program.cs
@@ -1,14 +1,0 @@
-﻿namespace Worker
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var builder = Host.CreateApplicationBuilder(args);
-            builder.Services.AddHostedService<Worker>();
-
-            var host = builder.Build();
-            host.Run();
-        }
-    }
-}
