--- src/LucasBenHur.MotorcycleMaintenance.Worker/Properties/launchSettings.json
@@ -1,15 +1,0 @@
-﻿{
-  "profiles": {
-    "Worker": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "DOTNET_ENVIRONMENT": "Development"
-      },
-      "dotnetRunMessages": true
-    },
-    "Container (Dockerfile)": {
-      "commandName": "Docker"
-    }
-  },
-  "$schema": "http://json.schemastore.org/launchsettings.json"
-}
