--- src/LucasBenHur.MotorcycleMaintenance.Worker/appsettings.json
@@ -1,8 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  }
-}
