+++ src/Services/RentService/Program.cs
@@ -1,0 +1,30 @@
+namespace Rent
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            var builder = WebApplication.CreateBuilder(args);
+
+            builder.Services.AddControllers();
+            builder.Services.AddEndpointsApiExplorer();
+            builder.Services.AddSwaggerGen();
+
+            var app = builder.Build();
+
+            app.UseSwagger();
+            app.UseSwaggerUI();
+
+            app.UseHttpsRedirection();
+
+            app.UseAuthorization();
+
+            app.MapControllers();
+
+            if (!app.Environment.IsDevelopment())
+                app.Run("http://*:80");
+
+            app.Run();
+        }
+    }
+}
